feat: modularized web wallet
- Closes #179
