+++ cmd/user-agent/src/pages/chapi/wallet/index.js
@@ -1,0 +1,11 @@
+/*
+Copyright SecureKey Technologies Inc. All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+export * from './common/util.js';
+export {DIDStore} from './common/didMetadata.js';
+export {WalletStore} from './store/saveCredential.js';
+export {WalletGet} from './get/getCredentials.js';
+export {DIDAuth} from './get/didAuth.js';
